+++ src/java/SavingsTransaction.xml
@@ -1,13 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN" "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
 <hibernate-mapping>
-    <class name="transactionpkg.SavingsTransaction" table="savingTrans">
-        <id name="transID" type="integer">
-            <generator class="increment"/>
-        </id>
-        <property name="acctid"/>
-        <property name="time"/>
-        <property name="location"/>
-        <property name="balance"/>
-        <property name="debitnumber"/>
-    </class>
+  <class name="com.syntelinc.BOK.ATM.transactionpkg.SavingsTransaction" table="savingTrans">
+    <id name="transID" type="integer">
+      <generator class="increment"/>
+    </id>
+    <property name="acctid"/>
+    <property name="time"/>
+    <property name="location"/>
+    <property name="balance"/>
+    <property name="debitnumber"/>
+  </class>
 </hibernate-mapping>
