+++ src/java/com/syntelinc/BOK/ATM/transactionpkg/SavingsTransaction.java
@@ -4,7 +4,7 @@
  * and open the template in the editor.
  */
 
-package transactionpkg;
+package com.syntelinc.BOK.ATM.transactionpkg;
 
 /**
  *
